:bug:(core): Drop fewer notifications
After converting notifications to a ring buffer, we were dropping them
on notification floods, even for clients that were consuming
notifications.  The buffer now drops old notifications after a 200ms
delay to give clients some time to catch up, and notifications are
processed in their own routine so that this delay doesn't block normal
responses.  The trade-off is that we lose notification ordering
guarantees, but I think this is probably acceptable.